+++ src/Console/Commands/Make/MakeModule.php
@@ -94,13 +94,13 @@ class MakeModule extends Command
 		$this->setUpStyles();
 		
 		$this->newLine();
-		
-		$this->ensureModulesDirectoryExists();
-		
+
 		if ($this->shouldAbortToPublishConfig()) {
 			return 0;
 		}
-		
+
+		$this->ensureModulesDirectoryExists();
+
 		$this->writeStubs();
 		$this->updateCoreComposerConfig();