+++ tests/testbench-core/app-modules/test-module/database/migrations/2024_04_03_133130_set_up_test-module_module.php
@@ -1,0 +1,13 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+
+return new class() extends Migration {
+	public function up()
+	{
+	}
+	
+	public function down()
+	{
+	}
+};